refactor: rearrange the scenes to scene.cpp/scene.h
