+++ include/sphere.h
@@ -48,7 +48,7 @@ private:
 	Material::Ptr mat_ptr;
 };
 
-bool Sphere::hit(
+inline bool Sphere::hit(
 	const ray& r, double t_min, double t_max, hit_record& rec) const
 {
 	vec3 oc = r.origin() - center;
